Factor some use of glob files
This changes the numbers in get_require_locations by 1, but I think that
function was previously unused.
